requirements/dev.txt
-r ./common.txt

django-extensions==3.2.1

# Debugging
ipdb==0.13.11

# API endpoints
httpie==3.2.1